component folder created and moved into it
