+++ components/categories/index.tsx
